engine/requirements.txt
# Core dependencies for Lokal Video Processing Engine
opencv-python==4.8.1.78
ultralytics==8.0.196
openai==1.3.7
supabase==2.0.2
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0

# Optional: For better performance
torch==2.1.0
torchvision==0.16.0

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0 